ClusterjConnector: disconnection and reconnection
Enable the NDB data access layer to detect disconnection events.
ClusterjConnectors are no longer responsible for keeping a per-thread
session cache.
ClusterjConnector(s) are now created by a ClusterjConnectorPool class
which keeps a per-thread cache of connectors and supports detection of
disconnection and reconnection.
The result of this is that single connector only contains a single
session which can be used by a thread.
To destroy a session, a connector can be marked as tainted and will be
discarded by the pool.